refactor: リポジトリを CQRS パターンに分離
既存コードベース (yantene.net.old) の実装方針に合わせ、
単一リポジトリを Command/Query に分離する。

- IStoredObjectMetadataQueryRepository: findAll, findByObjectKey
- IStoredObjectMetadataCommandRepository: upsert, deleteByObjectKey, incrementDownloadCount
- StoredObjectMetadataQueryRepository / StoredObjectMetadataCommandRepository を新規作成
- 旧 StoredObjectMetadataRepository を削除
- SyncService, 各ハンドラー, テストを新しいインターフェースに対応
