+++ app/backend/handlers/api/admin/files/index.ts
@@ -1,6 +1,7 @@
 import { drizzle } from "drizzle-orm/d1";
 import { Hono } from "hono";
-import { StoredObjectMetadataRepository } from "../../../../infra/d1/stored-object/stored-object-metadata.repository";
+import { StoredObjectMetadataCommandRepository } from "../../../../infra/d1/stored-object/stored-object-metadata.command-repository";
+import { StoredObjectMetadataQueryRepository } from "../../../../infra/d1/stored-object/stored-object-metadata.query-repository";
 import { StoredObjectStorage } from "../../../../infra/r2/stored-object.storage";
 import { SyncService } from "../../../../services/sync.service";
 import type { ErrorResponse, SyncResponse } from "~/lib/types/object-storage";
@@ -10,9 +11,17 @@ export const adminFilesApp = new Hono<{ Bindings: Env }>().post(
   async (c): Promise<Response> => {
     try {
       const db = drizzle(c.env.D1);
-      const repository = new StoredObjectMetadataRepository(db);
+      const queryRepository = new StoredObjectMetadataQueryRepository(db);
+      const commandRepository = new StoredObjectMetadataCommandRepository(
+        db,
+        queryRepository,
+      );
       const storage = new StoredObjectStorage(c.env.R2);
-      const syncService = new SyncService(storage, repository);
+      const syncService = new SyncService(
+        storage,
+        queryRepository,
+        commandRepository,
+      );
 
       const result = await syncService.execute();
